feat(api+e2e): Practice XP persistence + regression test suite
Full-stack fix for the user-reported bug where answering Practice
questions left the Home tier banner at 0/1000 forever. Five layered
issues blocked XP persistence — each one masked the next.

Backend (apps/api):

- SessionService.submitAnswer: restore creditNonRankedProgress call +
  method body. The method credits UserDailyProgress.pointsCounted for
  non-Ranked modes (Practice, daily, weekly, mystery, speed). Ranked
  keeps its own sync-progress path. Practice XP is capped at Tier 2
  (totalPoints >= 1000) — past that, Ranked drives progression.
  questionsCounted still ticks above the cap so streak / daily
  missions keep working.

- SessionService.submitAnswer: two-step user lookup (findById ||
  findByEmail). JWT subject is email for both mobile and OAuth
  logins — findById alone threw NoSuchElementException on every
  authenticated submission, which was only observable after the
  earlier DTO alias fix unblocked body parsing.

- UserQuestionHistoryRepository.deleteAllByUserId: change return
  type from `long` to `int`. Spring Data's @Modifying DML contract
  accepts only void/int/Integer; `long` threw
  InvalidDataAccessApiUsageException from admin reset-history during
  e2e setup.

- AdminTestController @Profile: add "docker". The test-fixture
  endpoints now load when SPRING_PROFILES_ACTIVE=docker so e2e
  suites targeting the docker-compose stack can seed state. The
  profile list still excludes production.

Tests (apps/api/src/test):

- SessionServiceTest: replace the stale
  submitAnswer_practiceMode_doesNotWriteUserDailyProgress case
  (asserted the opposite of the new contract) with two new cases:
  tier1_creditsXpToUserDailyProgress and tier2_capsXpButStillTicks
  Questions. Both wire up userTierService.getTotalPoints so the
  branch is exercised.

E2E (apps/web):

- W-M03-practice-xp.spec.ts: three-case suite pinning the contract.
  Pure-API flow (no UI) keeps the tests stable against the
  question-rendering critical path. L2-014 verifies tier-1 XP lands
  in /api/me/tier-progress; L2-015 verifies the Tier-2 cap; L2-016
  is a regression pin for the clientElapsedMs @JsonAlias fix.

All three pass against PLAYWRIGHT_BASE_URL=http://localhost:3000
with the docker-compose stack. SessionServiceTest + SessionController
Test 32/32 green.

+++ apps/api/src/main/java/com/biblequiz/modules/quiz/repository/UserQuestionHistoryRepository.java
@@ -45,7 +45,10 @@ public interface UserQuestionHistoryRepository extends JpaRepository<UserQuestio
            "GROUP BY h.question.book")
     List<Object[]> getAccuracyByBook(@Param("userId") String userId);
 
+    // Spring Data JPA @Modifying DML contract accepts only void / int /
+    // Integer return types. `long` threw InvalidDataAccessApiUsageException
+    // at call time (surfaced via admin reset-history during e2e setup).
     @org.springframework.data.jpa.repository.Modifying
     @Query("DELETE FROM UserQuestionHistory h WHERE h.user.id = :userId")
-    long deleteAllByUserId(@Param("userId") String userId);
+    int deleteAllByUserId(@Param("userId") String userId);
 }
